+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Add BBB backend app
 - Add BBB frontend app
 - Add form to create a scheduled video on the Dashboard
+- Add frontend component to register an email for a schedule webinar
 
 ### Changed